Refactor. Better UI and better everything
